File to set configuration for 3 screens
I've set a lot of options during time, just remove and set as you wish.

It can become much more advanced. but as I don't change this set as often, setting it in a fixed mode is fine until now. as it's a desktop, I don't change as often as a laptop would require.

I've set it to run in my i3wm setup